fix: include list item key in the view path of the list elements
